[GridNG] Fix computed style values for non-layout subgrids
This CL makes a few updates in order to round-trip specified values when
`subgrid` is specified on a non-subgrid.

Since subgrid tracks don't have sizes, we need to look at the track list
in various places instead of track sizes.

The concept of a collapsed auto repeater doesn't apply (since no
repetitions were computed), so I added an extra parameter to the
collectors to account for this.

grid-template-computed-nogrid.html didn't have any cases for integer
repeats (which behave differently than auto repeats, as they get
expanded in computed style), so I added a bunch, both with and without
auto repeats.

Bug: 618969
Change-Id: If348f6b099a9032f067fa2299c9d3444d2f52a26