view-transitions: Clip snapshots to max texture size.
If an element's painting exceeds max texture size, clip it down to max
texture size by painting the subset which is closest to the snapshot
root. This includes plumbing this subset to CC via paint clip nodes.
And setting the correct object-view-box + intrinsic size on the
replaced elements displaying the snapshot.

R=pdr@chromium.org, vmpstr@chromium.org, bokan@chromium.org

Bug:1409713
Change-Id: If81bc67ad6ffe49d68e4470a6717b704edafcd3f