[@scope] Enable nesting features within @scope's body
This CL passes a parent rule and accompanying CSSNestingType
to the ConsumeRuleList call which handles @scope's body. This has
the following effects:

 - A :scope descendant compound is implicitly inserted if the selector
   is not scope-containing.
 - The parent selector (&) now refers to the <scope-start>
   selector list.
 - Relative selectors are enabled.

The code to resize the arena in ConsumeStyleRule must now be detached
from whether or not parent_rule_for_nesting is nullptr. This is because
we can now enter ConsumeStyleRule with parent_rule_for_nesting!=nullptr
without another call to ConsumeStyleRule below us on the stack
(via ConsumeScopeRule).

Bug: 1280240
Change-Id: Ibe6b1e40e22d535a094b0ec464ae119952b015f5