Improve interaction of pointer-events with scroll hit testing and scrollability decisions
First, this change stops considering pointer-events when deciding
whether a paint layer is scrollable.  This avoids pointer-events having
strange effects on decisions about whether to show overlay scrollbars.
These strange effects include the backdrop of a <dialog> causing the
things behind it to become inert (and thus have pointer-events: none)
and cause overlay scrollbars to spontaneously appear.

This is done by changing a test of ComputedStyle::VisibleToHitTesting()
to test only ComputedStyle::Visibility(), which is one of the two parts
of VisibleToHitTesting().  This test dates back to
https://chromium.googlesource.com/chromium/src/+/9de09a320e201ea651b23587850050a67654271f%5E%21/
and I don't think the pointer-events part of the test makes sense.  (I'm
somewhat suspicious of the visibility part of the test as well, although
the authors of the tests at
web_tests/fast/scrolling/scrollable-area-{frame{,-scrolling-no}-visibility-hidden-child,overflow-auto-visibility-hidden{,-in-parent}}.html
believed in that part at least enough to write 4 tests for it.)  It
doesn't seem to me that pointer-events (or, really, visibility) should
affect the result of ScrollsOverflow since it doesn't affect whether
something can be scrolled programmatically (or, for pointer-events, via
the keyboard).

Additionally, this CL then undoes the effect that the above change would
have on decisions about implicit promotion to the root scroller.

On Mac (but not Linux), the two added tests show the bug without this
change, when run in virtual/overlay-scrollbar/, but are fixed by this
change.

Second, this change aligns the visibility tests used when constructing
hit test data used for compositor scrolling with the visibility tests
used by main thread hit testing, and forces the complicated case of a
scrollable region that is not visible to hit testing to use the main
thread hit testing codepath (since it may have arbitrary descendants,
scrollable or not, that are visible to hit testing).  These changes are
in box_painter.cc, with corresponding changes to scroll_area_painter.cc
to fix all relevant callers, and changes to paint_chunker.cc to adjust a
DCHECK.   The alignment aspect is needed to avoid regressing
external/wpt/pointerevents/pointerevent_hit_test_scroll.html on Mac
only.  Forcing the complex pointer-events cases to always fall back to
main thread hit testing is needed to avoid
pointerevent_hit_test_scroll_visible_descendant.html starting to fail
again on Linux and Windows as a result of the alignment aspect.  I don't
fully understand why the regression in pointerevent_hit_test_scroll.html
is Mac-only, but I know that on Linux, UpdateNonFastScrollableRegion
fails to bail out early in some cases where on Mac it does bail out
early because layer.element_id() == scroll_element_id; in these cases,
on Linux, layer.element_id() is 0 while scroll_element_id is not zero.

Third, this change also makes virtual/overlay-scrollbar/ and
virtual/non-overlay-scrollbar/ never expire, because they test something
that varies based on OS and settings rather than a feature being rolled
out (e.g., virtual/overlay-scrollbar/ matches the default behavior on a
Mac, whereas running outside that virtual suite does not), and renames
the README.txt to README.md to satisfy
https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/tools/blinkpy/web_tests/lint_test_expectations.py;l=345-348;drc=fdd1781f121a555a8f4b771b970147e1a7ec4876

Fixed: 1414142
Change-Id: Ic26cfb933948bca55e4042ed360ed07692d14a0e