Update ReadableStreamDefaultReader implementation
This change includes updating the ReadableStreamDefaultReader
implementation to match the Streams spec. This means properly implementing read requests so that they are not just
StreamPromiseResolvers, and have separate chunk, close, and error steps
based on where the read request is used. This CL also uses
ReadableStreamReadResult to match the spec, and finally removes the use
of `for_author_code_` which is no longer used in the standard. These
changes in turn also fix the teeing behavior.

Bug: 1227496
Change-Id: Id43ece0bef5fb5a7588e8adf64b5a4e398d5163e